+++ include/Hadouken/BitTorrent/TorrentInfo.hpp
@@ -31,7 +31,7 @@ namespace Hadouken
 
             int getNumPieces() const;
 
-            size_t getTotalSize() const;
+            uint64_t getTotalSize() const;
 
         private:
             std::shared_ptr<libtorrent::torrent_info> info_;
